+++ mithril-client/src/lib.rs
@@ -84,6 +84,7 @@ macro_rules! cfg_unstable {
 
 pub mod aggregator_client;
 cfg_unstable! {
+    pub mod cardano_stake_distribution_client;
     pub mod cardano_transaction_client;
 }
 pub mod certificate_client;